Template IPTagInfo.. do not expect it to work (some commented code here and there)
